refactor: 설문 문항, store 파일 위치 수정
