+++ src/app/ai/_constants/survey.ts
