+++ src/app/ai/coach/_store/financialGoalSurveyStore.ts
