+++ src/app/ai/coach/form/financial-goal/page.tsx
@@ -7,9 +7,9 @@ import StepController from '@/src/app/ai/coach/_components/form/StepController';
 import Step from '@/src/app/ai/coach/_components/form/Step';
 import Question from '@/src/app/ai/coach/_components/form/Question';
 import Answer from '@/src/app/ai/coach/_components/form/Answer';
-import { financialGoalSurvey } from '@/src/client/types/survey';
+import { financialGoalSurvey } from '@/src/app/ai/_constants/survey';
 import Modal from '@/src/app/common/Modal/Modal';
-import { useFinancialGoalSurveyStore } from '@/src/client/store/survey/financialGoalSurveyStore';
+import { useFinancialGoalSurveyStore } from '@/src/app/ai/coach/_store/financialGoalSurveyStore';
 import { useRouter } from 'next/navigation';
 import postFinancialGoalSurvey from '@/src/client/lib/api/postFinancialGoalSurvey';
 const totalStep = 12;
